+++ js/_website/src/routes/+layout.svelte
@@ -15,13 +15,12 @@
 	import { media_query } from "../utils";
 	store = media_query();
 
-	import { browser } from '$app/environment'
-    if (browser) {
-      window.__gradio_mode__ = "website";
-    }
+	import { browser } from "$app/environment";
+	if (browser) {
+		window.__gradio_mode__ = "website";
+	}
 	import version_json from "./version.json";
 	let version = version_json.version;
-
 </script>
 
 <svelte:head>
@@ -32,7 +31,8 @@
 
 	<script
 		async
-		src="https://www.googletagmanager.com/gtag/js?id=UA-156449732-1"></script>
+		src="https://www.googletagmanager.com/gtag/js?id=UA-156449732-1"
+	></script>
 	<script>
 		window.dataLayer = window.dataLayer || [];
 		function gtag() {
@@ -42,9 +42,7 @@
 		gtag("config", "UA-156449732-1");
 	</script>
 
-	<script
-		type="module"
-		src="https://gradio.s3-us-west-2.amazonaws.com/{version}/gradio.js"></script>
+	<script type="module" src="http://localhost:4321/gradio.js"></script>
 </svelte:head>
 
 <Header />