+++ js/gallery/src/Gallery.svelte
@@ -130,7 +130,8 @@
 		});
 	}
 
-	$: can_zoom = window_height >= client_height;
+	// If you don't allow preview, assume can always zoom to trigger select event
+	$: can_zoom = !allow_preview ? true : window_height >= client_height;
 
 	let client_height = 0;
 	let window_height = 0;