+++ client/python/CHANGELOG.md
@@ -16,6 +16,24 @@ No changes to highlight.
 
 No changes to highlight.
 
+# 0.2.7
+
+## New Features:
+
+- The output directory for files downloaded via the Client can now be set by the `output_dir` parameter in `Client` by [@abidlabs](https://github.com/abidlabs) in [PR 4501](https://github.com/gradio-app/gradio/pull/4501)
+
+## Bug Fixes:
+
+- The output directory for files downloaded via the Client are now set to a temporary directory by default (instead of the working directory in some cases) by [@abidlabs](https://github.com/abidlabs) in [PR 4501](https://github.com/gradio-app/gradio/pull/4501)
+
+## Breaking Changes:
+
+No changes to highlight.
+
+## Full Changelog:
+
+No changes to highlight.
+
 # 0.2.6
 
 ## New Features: